move eth and btc validators to a new submodule
